app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
pytz
SQLAlchemy
pymysql
cryptography